+++ src/main/java/org/zalando/nakadi/exceptions/runtime/TopicDeletionException.java
@@ -1,0 +1,8 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class TopicDeletionException extends MyNakadiRuntimeException1 {
+
+    public TopicDeletionException(final String msg, final Exception cause) {
+        super(msg, cause);
+    }
+}
